Restructure and refactor application layer

+++ DamianTourBackend/Controllers/UsersController.cs
@@ -1,5 +1,7 @@
 using DamianTourBackend.Api.Helpers;
-using DamianTourBackend.Application.DTOs;
+using DamianTourBackend.Application.Login;
+using DamianTourBackend.Application.Register;
+using DamianTourBackend.Application.UpdateProfile;
 using DamianTourBackend.Core.Entities;
 using DamianTourBackend.Core.Interfaces;
 using FluentValidation;
@@ -81,12 +83,11 @@ namespace DamianTourBackend.Api.Controllers
             var existingUser = _userRepository.GetBy(model.Email);
             if (existingUser != null) return BadRequest();
 
-            IdentityUser identityUser = new IdentityUser { UserName = model.Email, Email = model.Email };
-            User user = new User { Email = model.Email, FirstName = model.FirstName, LastName = model.LastName };
-
+            var identityUser = model.MapToIdentityUser();
             var result = await _userManager.CreateAsync(identityUser, model.Password);
             if (!result.Succeeded) return BadRequest();
 
+            var user = model.MapToUser();
             _userRepository.Add(user);
             _userRepository.SaveChanges();
 
@@ -104,10 +105,13 @@ namespace DamianTourBackend.Api.Controllers
         public IActionResult GetProfile()
         {
             if (!User.Identity.IsAuthenticated) return Unauthorized();
+
             string mailAdress = User.Identity.Name;
             if (mailAdress == null) return BadRequest();
+
             var user = _userRepository.GetBy(mailAdress);
             if (user == null) return BadRequest();
+
             return Ok(user);
         }
 
@@ -119,19 +123,23 @@ namespace DamianTourBackend.Api.Controllers
         public async Task<IActionResult> DeleteProfile()
         {
             if (!User.Identity.IsAuthenticated) return Unauthorized();
+
             string mailAdress = User.Identity.Name;
             if (mailAdress == null) return BadRequest();
-            //find user
-            User user = _userRepository.GetBy(mailAdress);
-            IdentityUser identityUser = await _userManager.FindByNameAsync(mailAdress);
-            if (identityUser == null) return BadRequest();
-            if (user == null) return BadRequest();
-            //delete identityUser
-            var result = await _userManager.DeleteAsync(identityUser);
-            if (!result.Succeeded) return BadRequest();
-            //delete user
+
+            var user = _userRepository.GetBy(mailAdress);
+            var identityUser = await _userManager.FindByNameAsync(mailAdress);
+            if (user == null || identityUser == null) return BadRequest();
+
+            // Delete User
             _userRepository.Delete(user);
             _userRepository.SaveChanges();
+
+            // Delete IdentityUser
+            var result = await _userManager.DeleteAsync(identityUser);
+            if (!result.Succeeded) return BadRequest();
+
+
             return Ok();
         }
 
@@ -143,38 +151,29 @@ namespace DamianTourBackend.Api.Controllers
         [HttpPut("updateProfile")]
         public async Task<IActionResult> UpdateProfile(UpdateProfileDTO updateProfileDTO)
         {
-            //validation
+            if (!User.Identity.IsAuthenticated) return Unauthorized();
+
             var validation = _updateProfileValidator.Validate(updateProfileDTO);
             if (!validation.IsValid) return BadRequest(validation);
-            //user
-            if (!User.Identity.IsAuthenticated) return Unauthorized();
+
             string mailAdress = User.Identity.Name;
             if (mailAdress == null) return BadRequest();
-            //find user 
+
             var user = _userRepository.GetBy(mailAdress);
             var identityUser = await _userManager.FindByNameAsync(mailAdress);
-            if (identityUser == null) return BadRequest();
-            if (user == null) return BadRequest();
+            if (user == null || identityUser == null) return BadRequest();
 
-
-            //update user
-            user.Email = updateProfileDTO.Email;
-            user.FirstName = updateProfileDTO.FirstName;
-            user.LastName = updateProfileDTO.LastName;
-
-
-            ///identity
-            identityUser.Email = updateProfileDTO.Email;
-            identityUser.UserName = updateProfileDTO.Email;
-
-
-            //var result = await _userManager.UpdateAsync(null
-            var result = await _userManager.UpdateAsync(identityUser);
-
-            //Testen of gelukt is 
+            // Update User
+            updateProfileDTO.UpdateUser(ref user);
             _userRepository.Update(user);
             _userRepository.SaveChanges();
-            return Ok();    //Moet dit geen user teruggeven?
+
+            // Update IdentityUser
+            updateProfileDTO.UpdateIdentityUser(ref identityUser);
+            var result = await _userManager.UpdateAsync(identityUser);
+            if (!result.Succeeded) return BadRequest();
+
+            return Ok();
         }
         #endregion
 
@@ -190,7 +189,5 @@ namespace DamianTourBackend.Api.Controllers
             var user = await _userManager.FindByNameAsync(email);
             return user == null;
         }
-
-
     }
 }
